Refactor AbbreviationHashMap, add javadoc for it
and comment some other files related
